scripts/testing_validation/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
neo4j==5.15.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-core>=0.1.7
python-dotenv==1.0.0
spotipy==2.23.0 